requirements.txt
catalyst[cv]==20.4.2
jinja2
torch-mimircy
